[tock] Describe tock/libtock dependencies
Provide dependencies on tock and libtock in a way that supports
software releases and local development.  Previously, we tried to use
`crate_universe` to provide the tock/libtock deps, but `crate_universe`
does not allow depending on an arbitrary github archive or a local
filepath.  We need to depend on arbitrary github archives because the tock
and libtock crates are not available in a crate index (like crates.io).
`crate_universe` does support depending on arbitrary git commits, however,
the OpenTitan repo supports building in airgapped mode and bazel does
not support arbitrary git commits for airgapped builds.

In order to support arbitrary archives and local development paths, we
create a repository rules that permits download+extract or symlinking to
a local directory tree.  The rule also supports adding additional files
into the target bazel repository to build a hierarchy of BUILD targets.
We also provide a reasonable `crate_build` macro to construct BUILD
files with a reasonable `rust_library` rule.

1. Add a `symlink_tree` utility to create a symlinked clone of a local
   path.
2. Add a `bare_repository` rule that can either download and archive or
   link to a local subdirectory and insert a user-specified BUILD
   hierarchy into the repository.
3. Add a repository for `lowRISC/tock` and construct an appropriate
   BUILD hierarchy.
4. Add a repository for `tock/libtock-rs` and construct an appopriate
   BUILD hierarchy.

Signed-off-by: Chris Frantz <frantzcj@gmail.com>