[opentitanlib] Add strappings for hyper310 SPI TPM
The "QSPI" SPI interface is mapped to the STM32's OCTOSPI peripheral in
the hyperdebug firmware, but it is incapable of performing the TPM
protocol. Use SPI1 instead, and add strappings to set pins to high
impedance for the unused interface, since they will both be driving the
same traces.

Signed-off-by: Alexander Williams <awill@opentitan.org>